pyproject.toml
[tool.poetry]
name = "prava-backend"
version = "0.1.0"
description = ""
authors = ["Matthias Wiench"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.5"
python-tcxparser = "^2.3.0"
python-multipart = "^0.0.17"
uvicorn = "^0.32.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
baml-py = "^0.68.0"
python-dotenv = "^1.0.1"
ipykernel = "^6.29.5"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
